requirements.txt
psycopg2-binary==2.9.10
python-dotenv==1.0.1
python-telegram-bot==20.0
asyncpg==0.29.0
